backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arcadegent-agent"
version = "0.1.0"
description = "Arcadegent FastAPI backend for arcade search and orchestration."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "fastmcp>=3.0.0,<4.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "pyyaml>=6.0.2,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
